# Conflicts:
#	flatlaf-core/build.gradle.kts
#	flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatComboBoxUI.java
#	flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatPasswordFieldUI.java

+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatArrowButton.java
@@ -48,8 +48,8 @@ public class FlatArrowButton
 	protected Color pressedBackground;
 
 	private int arrowWidth = DEFAULT_ARROW_WIDTH;
-	private int xOffset = 0;
-	private int yOffset = 0;
+	private float xOffset = 0;
+	private float yOffset = 0;
 
 	private boolean hover;
 	private boolean pressed;
@@ -126,19 +126,19 @@ public class FlatArrowButton
 		return pressed;
 	}
 
-	public int getXOffset() {
+	public float getXOffset() {
 		return xOffset;
 	}
 
-	public void setXOffset( int xOffset ) {
+	public void setXOffset( float xOffset ) {
 		this.xOffset = xOffset;
 	}
 
-	public int getYOffset() {
+	public float getYOffset() {
 		return yOffset;
 	}
 
-	public void setYOffset( int yOffset ) {
+	public void setYOffset( float yOffset ) {
 		this.yOffset = yOffset;
 	}
